+++ spec/helpers/link_helper_spec.rb
@@ -1,0 +1,32 @@
+require 'rails_helper'
+
+describe LinkHelper do
+  include IconHelper
+
+  # it 'foo' do
+    # user = create(:admin)
+    # puts link_to_edit([:backend, create(:admin)])
+  # end
+
+  # let(:model) { described_class }
+  # let(:klass) { model.to_s.underscore.to_sym }
+
+  # describe 'validations' do
+  #   it(:first_name) { expect(build(klass, first_name: nil)).not_to be_valid }
+  #   it(:last_name) { expect(build(klass, last_name: nil)).not_to be_valid }
+  #
+  #   describe 'email' do
+  #     it(:value) { expect(build(klass, email: nil)).not_to be_valid }
+  #
+  #     it :valid do
+  #       expect(build(klass, email: 'foo')).not_to be_valid
+  #       expect(build(klass, email: 'foo@bar.baz')).to be_valid
+  #     end
+  #
+  #     it :unique do
+  #       create(klass, email: 'foo@bar.baz')
+  #       expect(build(klass, email: 'FOO@BAR.baz')).not_to be_valid
+  #     end
+  #   end
+  # end
+end