+++ lib/udongo/cryptography.rb
@@ -1,0 +1,5 @@
+module Udongo::Cryptography
+  def crypt(options = {})
+    @crypt ||= Udongo::Crypt.new(options)
+  end
+end