+++ app/assets/javascripts/backend/general.js
@@ -1,6 +1,6 @@
 var general = general || {
   init: function() {
-    $('form input:visible:not(.no-focus):first').focus();
+    $('form:not(.no-focus) input:visible:not(.no-focus):first').focus();
   }
 };