+++ changelog.md
@@ -3,7 +3,7 @@
 * Add LinkHelper to provide shorthands for show/edit/destroy links.
 * Fix the sortable js to show the correct font-awesome icon.
 * Let Udongo::EmailVarsParser#replace_vars handle unlimited nested collections as var values.
-* Cleanup the layout of the email templates module.
+* Cleanup the layout of the following modules: email, email templates.
 
 
 0.0.13 - 2016-05-02