+++ changelog.md
@@ -7,6 +7,8 @@
   content.
 * Make it possible to disable the first input field focus for an entire form by
   adding the 'no-focus' class to that form.
+* Add Udongo::Cryptography to make encrypting/decrypting values secure and easy.
+  See readme.md for more information.
 
 
 0.0.13 - 2016-05-02