+++ changelog.md
@@ -3,8 +3,16 @@
 * Add LinkHelper to provide shorthands for show/edit/destroy links.
 * Add DatePickerInput to be used with Simple Form inputs (as: :date_picker)
 * Fix the sortable js to show the correct font-awesome icon.
-* Let Udongo::EmailVarsParser#replace_vars handle unlimited nested collections as var values.
-* Cleanup the layout of the following modules: email, email templates.
+* Let Udongo::EmailVarsParser#replace_vars handle unlimited nested collections
+  as var values.
+* Cleanup the layout of the following modules: email, email templates, flexible
+  content.
+* Make it possible to disable the first input field focus for an entire form by
+  adding the 'no-focus' class to that form.
+* Add Udongo::Cryptography to make encrypting/decrypting values secure and easy.
+  See readme.md for more information.
+* Make it possible to set the content column width for each responsive
+  breakpoint.
 
 
 0.0.13 - 2016-05-02