+++ changelog.md
@@ -15,6 +15,7 @@
   breakpoint.
 * Add the email validator and cleanup the URL validator.
 * Add DatePicker and DateRangerPicker to use with simple form.
+* Update to rails 4.2.6.
 
 
 0.0.13 - 2016-05-02